Refactor error messages messaging
