+++ packages/twenty-server/src/modules/connected-account/email-alias-manager/drivers/google/services/google-email-alias-manager.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@nestjs/common';
 
 import { google } from 'googleapis';
 
+import { GmailEmailAliasErrorHandlerService } from 'src/modules/connected-account/email-alias-manager/drivers/google/services/gmail-email-alias-error-handler.service';
 import { OAuth2ClientManagerService } from 'src/modules/connected-account/oauth2-client-manager/services/oauth2-client-manager.service';
 import { type ConnectedAccountWorkspaceEntity } from 'src/modules/connected-account/standard-objects/connected-account.workspace-entity';
 
@@ -9,6 +10,7 @@ import { type ConnectedAccountWorkspaceEntity } from 'src/modules/connected-acco
 export class GoogleEmailAliasManagerService {
   constructor(
     private readonly oAuth2ClientManagerService: OAuth2ClientManagerService,
+    private readonly gmailEmailAliasErrorHandlerService: GmailEmailAliasErrorHandlerService,
   ) {}
 
   public async getHandleAliases(
@@ -24,10 +26,14 @@ export class GoogleEmailAliasManagerService {
       auth: oAuth2Client,
     });
 
-    const emailsResponse = await peopleClient.people.get({
-      resourceName: 'people/me',
-      personFields: 'emailAddresses',
-    });
+    const emailsResponse = await peopleClient.people
+      .get({
+        resourceName: 'people/me',
+        personFields: 'emailAddresses',
+      })
+      .catch((error) => {
+        throw this.gmailEmailAliasErrorHandlerService.handleError(error);
+      });
 
     const emailAddresses = emailsResponse.data.emailAddresses;
